% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLike.R
\name{LogLikeMix}
\alias{LogLikeMix}
\title{Log-Likelihood estimation for mixed models}
\usage{
LogLikeMix(d, l, t, nu, r, c, SigSuppress, TwoSided)
}
\arguments{
\item{TwoSided}{}
}
\description{
Log-Likelihood estimation for mixed models
}
