% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamix-package.R
\docType{package}
\name{metamix-package}
\alias{metamix-package}
\alias{metamix}
\title{The 'metamix' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.36.0.9000. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
