% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamix.R
\name{metamix}
\alias{metamix}
\title{Meta-analysis Mixture Model}
\usage{
metamix(
  t,
  n1,
  n2 = NULL,
  SigOnly = FALSE,
  nonSigOnly = FALSE,
  alpha = 0.05,
  ConCof = 0.95,
  TwoSided = FALSE,
  SigSuppress = FALSE
)
}
\arguments{
\item{t}{vector of t-values}

\item{n1}{vector of sample sizes}

\item{n2}{second vector of sample sizes in case of two-sample t-tests}

\item{SigOnly}{did only significant study results enter the meta-analysis?}

\item{nonSigOnly}{did only non-significant study results enter the
meta-analysis?}

\item{alpha}{significance level (if one-sided: right tale)}

\item{ConCof}{confidence coefficient for calculating the confidence interval;
.95 corresponds to a 95% CI}

\item{TwoSided}{is the t-test two-sided?}

\item{SigSuppress}{if both significant and non-significant study results
entered the meta-analysis, were significant results suppressed?}
}
\description{
Estimate the true meta-analytical effect size corrected for publication bias.
}
