% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamixr-package.R
\docType{package}
\name{metamixr-package}
\alias{metamixr-package}
\alias{metamixr}
\title{metamixr: Bayesian Meta-Analytic Mixture Modelling}
\description{
Fits Bayesian random-effects meta-analytic mixture models in which the true study effects are drawn from a mixture of normal components with component-specific means and between-study heterogeneity. The mixture model can be combined with step-function selection models (Vevea and Hedges, 1995, \doi{10.1007/BF02294384}) to adjust for publication bias, either with one selection model for all studies or with separate selection models for groups of studies. Models are estimated by Hamiltonian Monte Carlo using 'rstan' and returned as 'stanfit' objects, so posterior summaries, posterior predictive checks and marginal likelihoods for model comparison (via 'bridgesampling') are readily available. Also includes a data simulator and the nudging meta-analysis data of Mertens et al. (2022) \doi{10.1073/pnas.2107346118}.
}
\details{
The package implements Bayesian meta-analytic mixture models with publication bias correction.
Three model fitting functions all return a
\code{\link[rstan:stanfit-class]{stanfit}} object:
\describe{
\item{\code{\link[=re_mix]{re_mix()}}}{random-effects meta-analytic mixture model without
publication bias adjustment (the ordinary random-effects meta-analysis
when \code{M = 1}).}
\item{\code{\link[=sel_mix]{sel_mix()}}}{the same mixture model combined with a step-function
selection model to adjust for publication bias.}
\item{\code{\link[=sel_flexpb]{sel_flexpb()}}}{the selection model but different selection processes
can be estimated for different groups of studies.}
}
\code{\link[=sim_mix]{sim_mix()}} simulates data from these models, and \link{mertens_nudge} contains
the nudging meta-analysis of Mertens et al. (2022) used in the vignettes.
}
\references{
Maier, M. (2026). Addressing heterogeneity with Bayesian meta-analytic
mixture modelling. \emph{PsyArXiv}. \doi{10.31234/osf.io/nkyqm_v2}

Maier, M., Bartoš, F., & Wagenmakers, E.-J. (2023). Robust Bayesian
meta-analysis: Addressing publication bias with model-averaging.
\emph{Psychological Methods}, 28(1), 107--122. \doi{10.1037/met0000405}

Stan Development Team (2024). RStan: the R interface to Stan.
\url{https://mc-stan.org/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MaxMaier42/metamix}
  \item \url{https://maxmaier42.github.io/metamix/}
  \item Report bugs at \url{https://github.com/MaxMaier42/metamix/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Maier \email{maximilianmaier0401@gmail.com} (\href{https://orcid.org/0000-0002-9873-6096}{ORCID})

Other contributors:
\itemize{
  \item Trustees of Columbia University [copyright holder]
}

}
\keyword{internal}
