% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixplot.R
\name{mixplot}
\alias{mixplot}
\title{Metamix plot}
\usage{
mixplot(x)
}
\arguments{
\item{x}{object of the class "metamix" as created by the `metamix` function}
}
\value{
ggplot object
}
\description{
Create an overview plot of the model-implied and empirical distribution of
t-values.
}
\examples{
# values from Table 2 in Ulrich et al. (2018)
n1_t2 <- c(20, 30, 35, 25, 60, 40, 45, 35, 70, 65,
           30, 40, 30, 20, 90, 70, 50, 65, 50, 45)
n2_t2 <- c(30, 35, 35, 20, 50, 40, 50, 30, 80, 60,
           25, 40, 25, 20, 80, 75, 50, 70, 50, 55)
t_t2 <- c(.59, 1.84, 1.72, -.4, .1, .48, 1.17, 2.18, .15, 1.76,
          1.9, .5, 1.27, 1.71, 1.17, 1.91, .08, -.48, .98, 2.28)
m_t2 <- metamix(t_t2, n1_t2, n2_t2)
mixplot(m_t2)
}
