% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.metamix.R
\name{print_metamix}
\alias{print_metamix}
\title{Print metamix output}
\usage{
print_metamix(x, doplot)
}
\arguments{
\item{x}{Object of the class "metamix" as created by the `metamix` function.}

\item{doplot}{should a metamix plot be shown as well?}
}
\description{
Print a summary of the metamix estimation results
}
