% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMix-package.r
\name{step3}
\alias{step3}
\title{Example output of parallel.temper() for use in the vignette/examples}
\format{A list with 2 elements}
\description{
Example output of parallel.temper() for use in the vignette/examples
}
