% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_mix.R
\name{re_mix}
\alias{re_mix}
\title{Random Effects Meta-Analytic Mixtures}
\usage{
re_mix(y, sd, M, mu_sd = 1, tau_sd = 0.2, ...)
}
\arguments{
\item{y}{vector of primary study effects}

\item{sd}{standard deviation of primary study effects}

\item{M}{number of mixture components}

\item{mu_sd}{standard deviation of prior on mean}

\item{tau_sd}{standard deviation of prior on heterogeneity}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Random Effects Meta-Analytic Mixtures
}
