% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_mix.R
\name{sel_mix}
\alias{sel_mix}
\title{Random Effects Meta-Analytic Mixtures with Selection Models}
\usage{
sel_mix(
  y,
  sd,
  M,
  steps = c(0.5, 0.95),
  one_sided = TRUE,
  mu_sd = 1,
  tau_sd = 0.2,
  ...
)
}
\arguments{
\item{y}{vector of primary study effects}

\item{sd}{standard deviation of primary study effects}

\item{M}{number of mixture components}

\item{steps}{p-value cutoffs (one-sided p-values). Currently only supports one or two steps.}

\item{one_sided}{whether selection is one or two-sided}

\item{mu_sd}{standard deviation of prior on mean}

\item{tau_sd}{standard deviation of prior on heterogeneity}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Random Effects Meta-Analytic Mixtures with Selection Models
}
