% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mix.R
\name{sim_mix}
\alias{sim_mix}
\title{Simulate Data from Random Effects Meta-Analytic Mixtures with Selection Models}
\usage{
sim_mix(
  K,
  M,
  mu,
  tau,
  steps,
  weights,
  one_sided,
  N_low = 25,
  N_high = 500,
  N_shape = 2,
  N_scale = 58
)
}
\arguments{
\item{K}{number of studies}

\item{M}{number of mixture components}

\item{mu}{vector of mixture means}

\item{tau}{vector of heterogeneity in each mixture component}

\item{steps}{p-value cutoffs (one-sided p-values). Currently only supports one or two steps.}

\item{weights}{relative publication probabilities}

\item{one_sided}{whether selection is one or two-sided}

\item{N_low}{lower bound on primary study sample sizes}

\item{N_high}{upper bound on primary study sample sizes}

\item{N_shape}{shape of negative binomial distribution to generate sample sizes (see Maier et al., 2023)}

\item{N_scale}{scale of negative binomial distribution to generate sample sizes (see Maier et al., 2023)}
}
\value{
A data frame with effect sizes and standard errors
}
\description{
Simulate Data from Random Effects Meta-Analytic Mixtures with Selection Models
}
